Find datastreams by tags not id - resolves #41
Datastream ids are now variable and include each device MAC code. This
means we cannot find them simply based on the id, instead we need to
detect the correct datastream by inspecting the tag contents to find the
appropriate match.

This commit also adds a spec helper which returns some 'real' JSON as
currently being published to Cosm be an Egg, and updates the dashboard
spec so that we test for the presence of the detected values.